Resolve "CTA frontend accepts files larger than a specified size"

+++ frontend/common/FrontendService.cpp
@@ -173,7 +173,7 @@ FrontendService::FrontendService(const std::string& configFilename) : m_archiveF
   // Initialise the Frontend
   auto archiveFileMaxSize = config.getOptionValueInt("cta.archivefile.max_size_gb");
   // Convert archiveFileMaxSize from GB to bytes
-  m_archiveFileMaxSize = archiveFileMaxSize.has_value() ?  archiveFileMaxSize.value()*1024*1024*1024 : 0;
+  m_archiveFileMaxSize = archiveFileMaxSize.has_value() ?  static_cast<uint64_t>(archiveFileMaxSize.value()) * 1024 * 1024 * 1024 : 0;
 
   {
     // Log cta.archivefile.max_size_gb
